try to let wordpress 300 sec before scanning
